chore: removed uneeded code from tests and source code
