--- setup-test-data/mock-destination-files/innerTwo.json
@@ -1,3 +1,0 @@
-{
-  "all": "~/.config/nested/again"
-}
